+++ flextype/Filters.php
@@ -1,4 +1,4 @@
-<?php namespace Flextype;
+<?php
 
 /**
  * @package Flextype
@@ -9,6 +9,8 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
+namespace Flextype;
 
 class Filters
 {
@@ -39,10 +41,6 @@ class Filters
 
     /**
      * Dispatch filters
-     *
-     *  <code>
-     *      Filter::dispatch('content', $content);
-     *  </code>
      *
      * @access  public
      * @param  string $filter_name The name of the filter hook.
@@ -82,14 +80,6 @@ class Filters
 
     /**
      * Add filter
-     *
-     *  <code>
-     *      Filter::add('content', 'replacer');
-     *
-     *      function replacer($content) {
-     *          return preg_replace(array('/\[b\](.*?)\[\/b\]/ms'), array('<strong>\1</strong>'), $content);
-     *      }
-     *  </code>
      *
      * @access  public
      * @param  string  $filter_name     The name of the filter to hook the $function_to_add to.