+++ flextype/I18n.php
@@ -1,6 +1,4 @@
-<?php namespace Flextype;
-
-use Symfony\Component\Yaml\Yaml;
+<?php
 
 /**
  * @package Flextype
@@ -11,6 +9,10 @@ use Symfony\Component\Yaml\Yaml;
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
+namespace Flextype;
+
+use Symfony\Component\Yaml\Yaml;
 
 class I18n
 {
@@ -130,10 +132,6 @@ class I18n
 
     /**
      * Initialize Flextype I18n
-     *
-     *  <code>
-     *      I18n::init();
-     *  </code>
      *
      * @access public
      * @return object