+++ flextype/Plugins.php
@@ -1,6 +1,4 @@
-<?php namespace Flextype;
-
-use Symfony\Component\Yaml\Yaml;
+<?php
 
 /**
  * @package Flextype
@@ -11,6 +9,10 @@ use Symfony\Component\Yaml\Yaml;
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
+namespace Flextype;
+
+use Symfony\Component\Yaml\Yaml;
 
 class Plugins
 {
@@ -43,7 +45,7 @@ class Plugins
 
             // Go through...
             foreach ($plugins_list as $plugin) {
-                if (Flextype::$filesystem->exists($_plugin = PLUGINS_PATH . '/' . $plugin . '/' . $plugin . '.yml')) {
+                if (Flextype::filesystem()->exists($_plugin = PLUGINS_PATH . '/' . $plugin . '/' . $plugin . '.yml')) {
                     $plugins_cache_id .= filemtime($_plugin);
                 }
             }
@@ -63,7 +65,7 @@ class Plugins
                 // Go through...
                 foreach ($plugins_list as $plugin) {
 
-                    if (Flextype::$filesystem->exists($_plugin_manifest = PLUGINS_PATH . '/' . $plugin . '/' . $plugin . '.yml')) {
+                    if (Flextype::filesystem()->exists($_plugin_manifest = PLUGINS_PATH . '/' . $plugin . '/' . $plugin . '.yml')) {
                         $plugin_manifest = Yaml::parseFile($_plugin_manifest);
                     }
 
@@ -89,10 +91,6 @@ class Plugins
 
     /**
      * Initialize Flextype Plugins
-     *
-     *  <code>
-     *      Plugins::init();
-     *  </code>
      *
      * @access public
      * @return object