+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+# Flextype 0.2.0, 2018-03-23
+* Thunderer Shortcode Framework - added
+* Cache Flextype::VERSION for cache key - added
+* flextype/boot/shortcodes.php	- removed
+* flextype/boot/events.php - removed
+* Code cleanup and refactoring #5
+
 # Flextype 0.1.0, 2018-03-21
 * Initial Release